#36572 新規登録画面で「削除」ボタンがある

+++ SRC/libArticle.php
@@ -165,7 +165,6 @@ function subArticleEdit()
 	$sPage   = $_REQUEST['sPage'];
 
 	$articleNo = $_REQUEST['articleNo'];
-
 	if ($articleNo) {
 		$sql = fnSqlArticleEdit($articleNo);
 		$res = mysqli_query($conn, $sql);
@@ -261,7 +260,11 @@ function subArticleEdit()
 
 		<a href="javascript:fnArticleEditCheck();"><img src="./images/<?php print $btnImage ?>" /></a>　
 		<a href="javascript:form.act.value='fManager';form.submit();"><img src="./images/btn_return.png" /></a>
-		&nbsp;&nbsp;<a href="javascript:fnArticleDeleteCheck(<?php print $articleNo ?>);"><img src="./images/btn_del.png" /></a>
+		<?php if ($articleNo) : ?>
+
+			&nbsp;&nbsp;<a href="javascript:fnArticleDeleteCheck(<?php print $articleNo ?>);"><img src="./images/btn_del.png" /></a>
+
+		<?php endif ?>
 	</form>
 <?php
 }
